Day 5/Task 1/requirements.txt
streamlit==1.45.0
google-generativeai==0.3.2
